Fix mapping TensorProto to NumPy for bfloat16 (#4234) (#4244)
* fix mapping for bfloat16


* fix typo


* remove old comments


* use storage_type for expected size instead


* nit: add comments for maps


* better comments for maps

(cherry picked from commit c2825db5276419cf74c0f23dd3dea8fd5ef7adad)

+++ onnx/mapping.py
@@ -3,6 +3,7 @@
 from onnx import TensorProto, SequenceProto, OptionalProto
 import numpy as np  # type: ignore
 
+# This map is used for converting TensorProto values into Numpy arrays
 TENSOR_TYPE_TO_NP_TYPE = {
     int(TensorProto.FLOAT): np.dtype('float32'),
     int(TensorProto.UINT8): np.dtype('uint8'),
@@ -13,7 +14,7 @@ TENSOR_TYPE_TO_NP_TYPE = {
     int(TensorProto.INT64): np.dtype('int64'),
     int(TensorProto.BOOL): np.dtype('bool'),
     int(TensorProto.FLOAT16): np.dtype('float16'),
-    int(TensorProto.BFLOAT16): np.dtype('uint16'),  # native numpy does not support bfloat16
+    int(TensorProto.BFLOAT16): np.dtype('float32'),  # Native numpy does not support bfloat16 so now use float32 for bf16 values
     int(TensorProto.DOUBLE): np.dtype('float64'),
     int(TensorProto.COMPLEX64): np.dtype('complex64'),
     int(TensorProto.COMPLEX128): np.dtype('complex128'),
@@ -23,9 +24,10 @@ TENSOR_TYPE_TO_NP_TYPE = {
 }
 
 # Currently native numpy does not support bfloat16 so TensorProto.BFLOAT16 is ignored for now
-# Numpy float16 array is only reversed to TensorProto.FLOAT16
+# Numpy float32 array is only reversed to TensorProto.FLOAT
 NP_TYPE_TO_TENSOR_TYPE = {v: k for k, v in TENSOR_TYPE_TO_NP_TYPE.items() if k != TensorProto.BFLOAT16}
 
+# This map indicates what storage-type is used in the protobuf (serialized) representation for TensorProto
 TENSOR_TYPE_TO_STORAGE_TENSOR_TYPE = {
     int(TensorProto.FLOAT): int(TensorProto.FLOAT),
     int(TensorProto.UINT8): int(TensorProto.INT32),
